requirements.txt
requests~=2.31.0
pandas~=2.2.1
typer~=0.12.0
python-dotenv~=1.0.1
rich~=13.7.1
zenpy~=2.0.47
